renamed the header files to .hpp
