+++ vesper_log/LoggingType.hpp
